refactor(hooks): introduce vueuse, delete duplicate hooks
