+++ src/hooks/event/useBreakpoint.ts
@@ -1,10 +1,10 @@
-import { ref, computed, Ref, unref } from 'vue';
-import { useEvent } from './useEvent';
+import { ref, computed, ComputedRef, unref } from 'vue';
+import { useEventListener } from '/@/hooks/event/useEventListener';
 import { screenMap, sizeEnum, screenEnum } from '/@/enums/breakpointEnum';
 
-let globalScreenRef: Ref<sizeEnum | undefined>;
-let globalWidthRef: Ref<number>;
-let globalRealWidthRef: Ref<number>;
+let globalScreenRef: ComputedRef<sizeEnum | undefined>;
+let globalWidthRef: ComputedRef<number>;
+let globalRealWidthRef: ComputedRef<number>;
 
 export function useBreakpoint() {
   return {
@@ -43,7 +43,7 @@ export function createBreakpointListen(fn?: Fn) {
     realWidthRef.value = width;
   }
 
-  useEvent({
+  useEventListener({
     el: window,
     name: 'resize',
     listener: () => {
