--- src/hooks/event/useEventHub.ts
@@ -1,10 +1,0 @@
-import { tryOnUnmounted } from '/@/utils/helper/vueHelper';
-import EventHub from '/@/utils/eventHub';
-const eventHub = new EventHub();
-export function useEventHub(): EventHub {
-  tryOnUnmounted(() => {
-    eventHub.clear();
-  });
-
-  return eventHub;
-}
