+++ src/hooks/event/useWindowSizeFn.ts
@@ -1,5 +1,4 @@
 import { tryOnMounted, tryOnUnmounted } from '/@/utils/helper/vueHelper';
-import { ref } from 'vue';
 
 import { useDebounce } from '/@/hooks/core/useDebounce';
 
@@ -37,22 +36,3 @@ export function useWindowSizeFn<T>(fn: Fn<T>, wait = 150, options?: WindowSizeOp
   });
   return [start, stop];
 }
-
-export const useWindowSize = (wait = 150, options?: WindowSizeOptions) => {
-  const widthRef = ref(0);
-  const heightRef = ref(0);
-
-  function setSize() {
-    widthRef.value = window.innerWidth;
-    heightRef.value = window.innerHeight;
-  }
-  setSize();
-
-  const handler = () => {
-    setSize();
-  };
-
-  useWindowSizeFn(handler, wait, options);
-
-  return { widthRef: widthRef, heightRef: heightRef };
-};
