--- src/hooks/state/useGlobalState.ts
@@ -1,12 +1,0 @@
-import { reactive } from 'vue';
-
-export function createGlobalState<T extends object>(factory: () => T) {
-  let state: T;
-
-  return () => {
-    if (!state) {
-      state = reactive(factory()) as T;
-    }
-    return state;
-  };
-}
