--- src/hooks/web/useLocalStorage.ts
@@ -1,5 +1,0 @@
-import { createStorage } from '/@/utils/storage';
-
-export function useLocalStorage() {
-  return createStorage(localStorage);
-}
