--- src/hooks/web/useSessionStorage.ts
@@ -1,5 +1,0 @@
-import { createStorage } from '/@/utils/storage';
-
-export function useSessionStorage() {
-  return createStorage(sessionStorage);
-}
